pulled out filter function  into utils
